Remove ResultSummaryCsv related services, repositories, and models
Refactor CSVWriter related code
