--- src/main/java/com/revature/repositories/ResultSummaryCsvRepository.java
@@ -1,9 +1,0 @@
-package com.revature.repositories;
-
-import com.revature.models.ResultSummaryCsv;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ResultSummaryCsvRepository extends JpaRepository<ResultSummaryCsv, Integer> {
-}
