+++ src/main/resources/application.properties
@@ -4,7 +4,7 @@ server.port=8083
 spring.datasource.url = ${db_url}
 spring.datasource.username = ${postgres_username}
 spring.datasource.password = ${postgres_pw}
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.default_schema=docutests3
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
